Resolve case conflict
